Fix other linter errors in admin-panel
