+++ .eslintrc
@@ -24,7 +24,8 @@
         "packageDir": ["../../", "."]
       }
     ],
-    "jsx-a11y/href-no-hash": "off"
+    "jsx-a11y/href-no-hash": "off",
+    "react/destructuring-assignment": "off"
   },
   "overrides": [
     // TypeScript packages
@@ -169,7 +170,6 @@
         "no-unused-vars": "warn",
         "react/button-has-type": "warn",
         "react/default-props-match-prop-types": "warn",
-        "react/destructuring-assignment": "off",
         "react/no-unused-prop-types": "warn",
         "react/no-array-index-key": "warn",
         "react/forbid-prop-types": "warn",
