+++ packages/admin-panel/src/api/TupaiaApi.js
@@ -134,6 +134,6 @@ const createTimeoutPromise = () =>
   new Promise((resolve, reject) => {
     const id = setTimeout(() => {
       clearTimeout(id);
-      reject({ message: 'Network request timed out' });
+      reject(new Error('Network request timed out'));
     }, FETCH_TIMEOUT);
   });
