+++ packages/admin-panel/src/pages/resources/ResourcePage.js
@@ -80,6 +80,7 @@ ResourcePage.propTypes = {
   title: PropTypes.string.isRequired,
   baseFilter: PropTypes.object,
   defaultSorting: PropTypes.array,
+  defaultFilters: PropTypes.array,
 };
 
 ResourcePage.defaultProps = {
@@ -91,5 +92,6 @@ ResourcePage.defaultProps = {
   onProcessDataForSave: null,
   baseFilter: {},
   defaultSorting: [],
+  defaultFilters: [],
   reduxId: null,
 };
