+++ packages/admin-panel/src/table/selectors.js
@@ -3,7 +3,7 @@
  * Copyright (c) 2018 Beyond Essential Systems Pty Ltd
  */
 
-import { DEFAULT_TABLE_STATE } from './reducer';
+import { DEFAULT_TABLE_STATE } from './constants';
 
 export const getTableState = (state, reduxId) => state.tables[reduxId] || DEFAULT_TABLE_STATE;
